[REFACTOR] Move Login form as a shared component between stages
